refactor: split deadline scheduling responsibilities to break dependency cycle
